+++ firmware/src/NativeTestMain.cpp
@@ -1,8 +1,0 @@
-#if defined(UNIT_TEST) || defined(PIO_UNIT_TESTING)
-
-int main() {
-    return 0;
-}
-
-#endif
-