+++ firmware/test/fakes/FakeWiFi.cpp
@@ -4,57 +4,62 @@
 
 #include <algorithm>
 
+namespace {
+    wl_status_t g_status = WL_IDLE_STATUS;
+    String g_connectedSsid;
+    String g_ipAddress = "0.0.0.0";
+    bool g_autoReconnect = false;
+    unsigned long g_millis = 0;
+
+    std::vector<wl_status_t> g_resultSequence;
+    std::size_t g_resultIndex = 0;
+    std::size_t g_runCallCount = 0;
+}
+
 FakeWiFiClass WiFi;
 FakeSerialClass Serial;
 
-std::vector<wl_status_t> FakeWiFiMulti::runResults;
-std::size_t FakeWiFiMulti::runResultIndex = 0;
-std::size_t FakeWiFiMulti::runCallCount = 0;
-
-namespace {
-unsigned long fakeMillisValue = 0;
-}
-
-unsigned long millis() {
-    return fakeMillisValue;
-}
+FakeWiFiMulti::FakeWiFiMulti() = default;
 
 FakeWiFiClass::FakeWiFiClass() {
     reset();
 }
 
 void FakeWiFiClass::reset() {
-    currentStatus = WL_IDLE_STATUS;
-    connectedSSID.clear();
-    ipAddress = "0.0.0.0";
+    g_status = WL_IDLE_STATUS;
+    g_connectedSsid.clear();
+    g_ipAddress = "0.0.0.0";
+    g_autoReconnect = false;
 }
 
 int FakeWiFiClass::mode(int) {
-    return 0;
+    return WIFI_STA;
 }
 
 void FakeWiFiClass::setHostname(const char*) {}
 
-void FakeWiFiClass::setAutoReconnect(bool) {}
+void FakeWiFiClass::setAutoReconnect(bool enable) {
+    g_autoReconnect = enable;
+}
 
 wl_status_t FakeWiFiClass::status() const {
-    return currentStatus;
+    return g_status;
 }
 
 IPAddress FakeWiFiClass::localIP() const {
-    return IPAddress(ipAddress);
+    return IPAddress(g_ipAddress);
 }
 
 String FakeWiFiClass::SSID() const {
-    return connectedSSID;
+    return g_connectedSsid;
 }
 
 String FakeWiFiClass::SSID(int) const {
-    return connectedSSID;
+    return g_connectedSsid;
 }
 
 int FakeWiFiClass::RSSI(int) const {
-    return -60;
+    return -55;
 }
 
 int FakeWiFiClass::scanNetworks(bool) {
@@ -68,61 +73,72 @@ int FakeWiFiClass::scanComplete() const {
 void FakeWiFiClass::scanDelete() {}
 
 void FakeWiFiClass::begin(const char*, const char*) {
-    currentStatus = WL_DISCONNECTED;
+    g_status = WL_DISCONNECTED;
 }
 
 void FakeWiFiClass::setStatus(wl_status_t status) {
-    currentStatus = status;
+    g_status = status;
 }
 
 void FakeWiFiClass::setConnectedSSID(const String& ssid) {
-    connectedSSID = ssid;
+    g_connectedSsid = ssid;
 }
 
 void FakeWiFiClass::setIPAddress(const String& ip) {
-    ipAddress = ip;
+    g_ipAddress = ip;
 }
-
-FakeWiFiMulti::FakeWiFiMulti() = default;
 
 void FakeWiFiMulti::addAP(const char* ssid, const char* password) {
     knownAPs.emplace_back(String(ssid ? ssid : ""), String(password ? password : ""));
 }
 
 wl_status_t FakeWiFiMulti::run(uint32_t) {
-    ++runCallCount;
-    wl_status_t result = (runResultIndex < runResults.size())
-        ? runResults[runResultIndex++]
-        : WiFi.status();
+    ++g_runCallCount;
 
-    WiFi.setStatus(result);
-    if (result == WL_CONNECTED && WiFi.SSID().empty() && !knownAPs.empty()) {
-        WiFi.setConnectedSSID(knownAPs.front().first);
+    wl_status_t result;
+    if (g_resultIndex < g_resultSequence.size()) {
+        result = g_resultSequence[g_resultIndex++];
+    } else {
+        result = g_status;
     }
+
+    if (result == WL_CONNECTED) {
+        g_status = WL_CONNECTED;
+        if (g_connectedSsid.empty()) {
+            if (!knownAPs.empty()) {
+                g_connectedSsid = knownAPs.front().first;
+            }
+        }
+    } else {
+        g_status = result;
+    }
+
     return result;
 }
 
 void FakeWiFiMulti::setRunResults(const std::vector<wl_status_t>& results) {
-    runResults = results;
-    runResultIndex = 0;
-    runCallCount = 0;
+    g_resultSequence = results;
+    g_resultIndex = 0;
+    g_runCallCount = 0;
 }
 
 void FakeWiFiMulti::resetResults() {
-    runResults.clear();
-    runResultIndex = 0;
-    runCallCount = 0;
+    g_resultSequence.clear();
+    g_resultIndex = 0;
+    g_runCallCount = 0;
 }
 
 std::size_t FakeWiFiMulti::getRunCallCount() {
-    return runCallCount;
+    return g_runCallCount;
 }
 
 namespace FakeWiFiShim {
     void reset() {
         WiFi.reset();
         FakeWiFiMulti::resetResults();
-        setMillis(0);
+        g_runCallCount = 0;
+        g_resultIndex = 0;
+        g_millis = 0;
     }
 
     void setInitialStatus(wl_status_t status) {
@@ -139,17 +155,20 @@ namespace FakeWiFiShim {
     }
 
     void setMillis(unsigned long value) {
-        fakeMillisValue = value;
+        g_millis = value;
     }
 
     unsigned long getMillis() {
-        return fakeMillisValue;
+        return g_millis;
     }
 
     std::size_t getRunCallCount() {
-        return FakeWiFiMulti::getRunCallCount();
+        return g_runCallCount;
     }
 }
 
-#endif
+unsigned long millis() {
+    return g_millis;
+}
 
+#endif